backend/requirements.txt
django

djangorestframework
markdown
django-filter

django-redis==5.2.0
celery==5.2.3
redis==4.1.3
python-dotenv==0.19.2
pydevd-pycharm==231.8109.197
django-celery-beat==2.2.1

psycopg2-binary

selenium
webdriver-manager
python-dateutil
xlsxwriter
